Fix build config and add simple workbench
